Looks for fact gathering and fix parser
We want to make sure we account for fact gathering.  Also, the
parser was not handling the first task correctly.  The way this
works is that the current task as associated with it how long it
took to run the last task, so on start there is no "last task".
There's probably a bug on the other end when we get to the last
task we are not looking at how long it took to run.  I'll fix
that another time :)